build: Disable no-unused-vars rule in TypeScript
Instead, rely on the typescript-eslint version of the rule, and upgrade
it from a warning to an error.

Annoyingly, we have to repeat these settings in the package-specific
eslintrc files too, because the wikimedia/client-es6 and
wikimedia/server rulesets turn no-unused-vars back on.

(Technically we only need to disable no-unused-vars in those files, we
don't have to re-enable @typescript-eslint/no-unused-vars, but doing so
explicitly makes it clearer that we're replacing the rule rather than
disabling it completely.)

Change-Id: I176edc61e9c07be27c6c93464cc880c34448078d