TextInput: Remove `dir="auto"` attribute for i18n
Although the HTML spec states that the directionality of a
directionality-capable attribute should be determined by the first
character in the attribute's value [1], at least Chrome, Safari, and
Firefox (probably others) don't seem to follow this [2] which can result
in placeholder text appearing on the wrong side of the input.

Because of this, setting the `dir` attribute on `input` elements
should be up to the client to decide (this is possible via
`v-bind="$attrs"`). Notably, Vector legacy's search input does not set a
dir attribute which is how it bypasses this issue.

Equivalent patch from WVUI at Ib98f2e7cdc.

[1] https://html.spec.whatwg.org/multipage/dom.html#directionality-of-the-attribute
[2] https://github.com/whatwg/html/issues/4903

Bug: T279674
Change-Id: I7b137a0db7b3652e572f1ac5bad1ae55fbb1cb04