Use a generic type for useModelWrapper's ModelValue
The type of the modelValue prop will vary per component, based on
the type of component, element, or input that is using that value.
Instead of trying to create a type that covers all cases, just use
a generic type called ModelValue which TypeScript will infer from
the type of the modelValue prop being passed in.

Change-Id: I68299219fcfd837f662d2ed3379828f13baa1bae