Radio: fix clickable area so it spans width of label text
Previously, `display: block` was set on the label element
so that groups of radios would be stacked vertically, but
this meant that blank space to the right of the text would
be clickable and would toggle the radio on click.

This patch wraps the label in a span, which is set to
`display: block`, so the label element is only as wide as
its text.

Change-Id: I95db620afea8726666e729c1b16a022900bef002