Select, Option: Introduce Select & Option components plus useMenu
The Select component takes in options as a prop in the form of an
array of MenuOption types. Scoped slots are provided to allow the
end-user to customize the display of the handle label and the menu
options. Basic usage looks like this:

    <cdx-select
        v-model="selection"
        :options="options"
        :default-label="'Choose an option'"
        :disabled="isDisabled"
    />

Internally, an Option component is used for each item in the menu,
but the end user does not need to provide it themselves.

In addition to the Select and Option components, this patch includes
a new composable called useMenu, which provides functionality and
helpers for components that contain menus. This composable includes:
- Adding a generated ID to each option
- A `state` object that contains which option is selected,
  highlighted, and/or active, for use in providing the required UX
- A ref for tracking menu visibility and a blur handler to close the
  menu
- A handleOptionChange method to make changing menu state easy and
  consistent
- Comprehensive keyboard navigation handling

Bug: T295506
Bug: T293302
Change-Id: Iaddc326e477353f2bcbe04e70a11ef2fe98bac79