build: Also watch design tokens in "npm run dev"
We were already doing this in npm run doc:dev because design tokens are
used and visualized in the documentation. But now that we're about to
start using tokens in components too, we also need it for "npm run dev",
so that changes to the design tokens immediately update in the
components sandbox.

Change-Id: Id1ca885698508720238c8e858f7b0a12c2d9f5c2