icons: Align non-default 'bold*' and 'italic*' language icons to guidelines
This is based on amendments to the WikimediaUI collection
https://www.figma.com/file/1lT9LKOK6wiHLnpraMjP3E/%E2%9D%96-WikimediaUI-%E2%80%93-Icons-and-Colors?node-id=1673%3A2052,
based on Design Style Guide's guideline following 'Noto Sans' usage.
This patch also brings icons in sync to OOUI v0.43.0.
The icons provided on task have been aligned to the Wikimedia SVG
guidelines with help of shared SVGO config. Additionally the
floating point precision was lowered in some cases and double-checked on
200% and 400% zoom in browser.

Bug: T298211
Change-Id: Ie17442e632b71d44db6c2222def79a3b47b2e8e0