build: Use UMD instead of CJS
UMD is very similar to CJS, but includes a little bit of extra wrapper
code that puts the library in the global scope (as window.codex) in
non-CJS environments.

This is needed to include Codex in MediaWiki due to T284511. Once
T284511 is fixed, the CJS build would have worked in MediaWiki, but the
UMD build will continue to work too. For better compatibility, we might
as well publish a UMD build rather than CJS.

Also change 'name' from 'Codex' to 'codex', so that our library ends up
at window.codex rather than window.Codex.

Bug: T299136
Change-Id: Id7fa8059e2f3936c2d725c1c4de41fd8913bc213