Lookup: listen to update:modelValue event instead of input events
The TextInput component emits input events only on input. It emits
update:modelValue events anytime the value of the input changes,
which is more valuable to a parent component like Lookup. This
patch changes the Lookup component to listen for update:modelValue,
specifically to ensure that it will emit a new-input event if the
input is cleared via the clear button.

This patch also changes update:inputValue to new-input to avoid
confusion with events emitted when a prop bound by v-model is
updated.

Change-Id: Ieeafd05d02dcd99133df4143759ac41ff38e1107