TypeaheadSearch: Add initial iteration of TypeaheadSearch
This patch contains:
- A generalized version of the TypeaheadSearch component from WVUI
- Generalized and renamed versions of the TypeaheadSuggestion and
  TypeaheadSuggestionTitle components from WVUI
- Some string helper functions from WVUI
- Updates to the useMenu composable: the composable now accepts an
  optional configuration object that can modify its behavior to
  suit the Typeahead use-case
- Demos of TypeaheadSearch used to search Wikipedia articles and
  Wikidata items
- Demos of the ListTile component, including important test cases
  from WVUI

This version of TypeaheadSearch focuses on providing the UI of the
component and does not take in a search client, a URL generator,
a domain, or a search page title. Instead, input events are
emitted up to the parent, which will fetch or compute search
results, then provide them to TypeaheadSearch via the suggestions
prop. These results must be pre-formatted according to the
SearchResult type.

Language has been changed to meet Codex norms and to remove any
references to MediaWiki or MediaWiki-specific content (e.g.
"title" has been changed to "label" since search results need
not represent articles with titles).

Menu behavior is mostly provided by the useMenu composable,
although some custom behavior is included in TypeaheadSearch.

Bug: T299680
Bug: T299682
Change-Id: I180ca46f89cd21712a2affa30ee4a8824d880ae6