build: Fix RTL flipping so that comment directives work
Our previous approach to RTL flipping was a Vite plugin that took the
built and minified codex.style.css and ran it through RTLCSS to produce
codex.style-rtl.css. However, that approach breaks comment directives
like /*rtl:ignore*/, because they are removed by the minifier.

Vite makes it very difficult to bifurcate CSS output or access
the unminified CSS output, but it does allow us to install PostCSS
plugins. This patch uses the following, hackier approach to generate RTL
output that does respect comment directives:

- Use the postcss-rtlcss plugin to transform CSS to bidirectional CSS,
  e.g. .foo { left: 0 } becomes
  [dir=ltr] .foo { left: 0 } [dir=rtl] .foo { right: 0 }
- In serve mode (npm run dev), use this bidirectional CSS unmodified
- In build mode (npm run build), use a custom PostCSS plugin to
  transform the bidirectional CSS back to LTR CSS before it's saved to
  codex-style.css. Also transform the bidirectional CSS to RTL CSS, and
  save that to codex-rtl.css.

The output of this approach is identical to the previous one, except for
the order of some CSS properties, and the names of animations (which now
end in -ltr or -rtl; renaming these back is possible, but not worth the
effort for now).

Bug: T295189
Change-Id: I8363fb2633b3758053a29d25cb9836ad70c07278