docs: Add direction switcher to docs
Add the postcss-rtlcss plugin in Vitepress too, but only apply it to
components CSS, leaving the CSS for the Vitepress UI itself alone. This
means the component demos flip when the direction is changed, not the
whole page.

Because PostCSS doesn't support this kind of directory scoping natively,
add a scoping plugin that wraps another PostCSS plugin and scopes it to
a certain set of file paths. Use the filter utility from
@rollup/pluginutils, so that the 'include' and 'exclude' settings work
exactly the same as they would for a Rollup or Vite plugin.

Add a direction switcher to the header, but only display it on component
pages. On other pages it doesn't do anything and would be confusing. We
should probably look into a better design for this, and with a smaller
design for the switcher we could consider putting a separate switcher on
each <Wrapper>.

In order to make the direction switcher work for icons, rerender the
every demo every time the direction is switched, by applying :key="dir".
To demonstrate this, change the icons used in the icon demos to
flippable ones.

(Add skipLibCheck=true to tsconfig, because importing Vitepress causes
annoying errors that haven't yet been fixed upstream.)

Bug: T296032
Change-Id: I060c627a4dd609e60104e61a050801968fda75e1