docs: Add documentation about bidirectionality
Also add documentation about which files are in the build.

This adds all the bidi documentation to the "Contributing code" section,
but that isn't the right place for all of it. Most of it is for people
contributing code to the library, but some of it is more relevant to
consumers of the library, so maybe this should be structured
differently.

Bug: T295195
Bug: T295196
Change-Id: I8d541d36f5236c6ef730e3790c43bec34fd28adc