build, icons: Make `npm test` fail if there are uncommitted SVG changes
As part of `npm test`, run the build and verify that there aren't any
uncommitted changes in the packages/icons/src/images/ directory. When
running on a clean git state (like in CI), this verifies that any
modifications to the SVG files caused by the build process (in
particular, the optimizations in `build:svg`) are committed to the
repository.

Also reorganize the `outputTypes` test into a separate `test:output`
command, and run the build as part of `npm test` (which the other
workspaces already do).

Change-Id: I7577ee9f846560a5eae40a765b6ae86dccd28978