vitepress: Fix inject warning in Wrapper test
The test wasn't failing, but it threw a big ugly warning about the
DirectionKey injection being missing. This happens because the tests
load the Wrapper component without the CustomLayout component, which is
what injects the direction.

Fix this by injecting a dummy direction ref (simply set to 'ltr') when
mounting the Wrapper component in the tests.

Change-Id: Ie286a43d20d9df7167b34d4c18192917a8605531