docs: Add warning about a TS pitfall with component template refs
Using "as typeof CdxFoo" is tempting because it seems right, but it's
actually wrong, and it's hard to notice because it doesn't cause errors.
The lack of errors is because the resulting type lets you access any
property it doesn't know about and get an "any" type back (summarized in
the docs as "disables almost all type checking").

Change-Id: Ic06121d36c674459956f77586e1ce7f38558c602