tokens: Add 'size-indicator'
'size-indicator' is part of existing components, and although
challenged in the past in T184022, we continue to need them for the
close future. Specifically as we're providing required and clear
indicating icons in input field components, we won't update those
icon sizes before a complete component review down the road after tokens
definitions.
Until then this token is added as is and not flagged deprecated.

Bug: T295711
Change-Id: I852975157cbe16da92b6084ee0aef5b6ac127b7b