build: Copy SVGs to dist/images at the right time
Other changes moved this step too early in the build process, so that
the icon SVGs were copied to dist/images, but were then deleted again by
the Vite build step. Change the order so that the SVGs are still
minified before the Vite build, but are copied to dist/images after the
Vite build.

Change-Id: I2a04856b713f646358f67b904074b33354f6361d