build: Enable safeBothPrefix for postcss-rtlcss
This prevents specificity bugs in the demo. Specifically, the
TypeaheadSearch patch has CSS that looks like this:

    .foo {
        border-right-color: transparent;
    }

    .foo:hover {
        border-color: #72777d;
    }

postcss-rtlcss changed the first rule to [dir=ltr] .foo, which made it
more specific than the hover rule, so the hover effect didn't work. To
fix this, safeBothPrefix changes the second rule to [dir] .foo:hover, so
that it's still more specific than the rule it's supposed to override.

Change-Id: Ic5f7f7f5d83b7da079736bebcb4279ab4385435b