build: Add 'branch-deploy' npm script, for WMF CI to call
This script builds the documentation and deploys it to Netlify when the
right environment variables are passed in. It's meant to be called by
WMF CI when a new change is uploaded to Gerrit.

Inspired by I37a0b4a82b894671989105230b41ebdf650be3e5 in
wikidata/query-builder.

The corresponding change in the WMF CI config is
I06a34fdec51b0cc673cee7189f69ba05ce645e63.

Bug: T293705
Change-Id: I82bb491ac69fdbd78ad413270c61a51e79772e43