docs: Standardize on cdx-docs prefix
Some components internal to the docs site only had one word names,
and the CSS class prefix we were using ("vp-") was unclear and
pointed to the technology we're using for our docs site (Vitepress)
instead of what the docs site is (Codex docs).

This patch renames all components internal to the docs site to
CdxDocs[ComponentName], and updates all CSS class names to be
prefixed with "cdx-docs-" instead of "vp-".

Bug: T302839
Change-Id: Id57028e949ccb0389b680b678f475e76466c991f