build: Require all CSS classes to start with `cdx-`
In the codex-docs package, require all classes to start with
`cdx-docs-`. This is a bit too strict because we sometimes do need to
style components in the docs CSS, but currently there's only one
instance of that.

Fix violations, including one that caused a small visual bug in
ToggleSwitch (the 1-pixel border on the grip didn't turn blue when
focused).

Bug: T302839
Change-Id: I2107cb753229f06c5306230bdc89fe8872291cbe