docs: Use kebab-case for component names in *.md files
Unfortunately we're not able to enforce this with eslint right now, so
this was a manual process, and new uses of PascalCase could creep in
over time.

Also remove unused imports in text-input.md. And fix code examples to
use self-closing tags where expected and once a missing `cdx-` prefix.

Change-Id: I7ac684c3992eb4d0d39aa6a21535e5f58fac99cc