MenuItem: Change Option to MenuItem
This patch changes the Option component to MenuItem to meet
design specs. Because this conflicts with the MenuItem type,
that type has been renamed to MenuItemData.

This will make more sense when the ListTile component is merged
into MenuItem, because MenuItem will have props not only
related to properties of the MenuItemData type, but additional
props related to display (e.g. highlightQuery).

Bug: T300625
Change-Id: I88302f440085839fac93c6df31279a1aefc0a693