docs: Set dir="ltr" on all non-component docs pages
To make bidirectional styles work, every Codex component in the docs has
to have exactly one ancestor with a dir attribute. On component pages we
manage this in the Wrapper component, but some non-component pages also
use components. The direction doesn't need to be switchable on those
pages, so just apply dir="ltr" to the root container.

This fixes the display of buttons on the icons/overview page, and
probably some other things we haven't found yet.

Change-Id: I80faff7978bde9b5615b7aef116d1b5d93479ec2