docs, Wrapper: add a "copy" button for code samples
When a component demo has a code sample to show, include a
button to copy the text of that sample. The button is only shown
when the code sample is toggled to visible, and appears right
next to the toggle, where it should be easy to notice.

The button itself is the existing CopyTextButton component,
previously under /components/tokens/ for use in copying
the names of tokens from the demonstration pages. The button
is moved to a dedicated folder /components/copy-text-button/
to make it clearer that it can be used for things other than just
token names.

Bug: T295184
Change-Id: I251f6dd92b4b1c5de29fdc035b91baaf82bcad3b