Menu, TypeaheadSearch: Remove selectHighlighted prop
For the TypeaheadSearch component, navigating between results via
keyboard should not cause the "selection" state to change, but the
input value should still be updated to the label or value of the
item that was navigated to.

Previously a "selectHighlighted" property existed on the Menu
component which allowed keyboard navigation to a given item to
update the select and highlight states at the same time. Since
this UX is not correct, this prop has been removed.

Instead, in order to continue supporting the "change input text
on keyboard nav" behavior that TypeaheadSearch needs, the Menu
component now emits a "menu-item-keyboard-navigation" event
with data for the relevant item as a payload. Most Menu-embedding
components (like Combobox or Select) can simply ignore this event.
TypeaheadSearch can listen for this event in order to update
its input value when necessary.

Bug: T303631
Change-Id: I0c26ae92dd5269dc4533404179ed7435c59b98e3