build: Add the @wikimedia/codex-search package
This package doesn't have any source code itself, it just exports the
TypeaheadSearch component from packages/codex. This provides a
slimmed-down build of Codex with only those things relevant for search
use.

Importing from '../codex/src/...' is ugly, but unfortunately required as
it's the only way to make sure that the CSS is tree-shaken. Importing
from '@wikimedia/codex-search' imports none of the CSS, and importing
from ../codex/src/lib.ts (aliased to '@wikimedia/codex-search') imports
all of the CSS in the entire library.

Reuse the Vite config from the codex package, since the only thing
that's different is the library name.

Bug: T303569
Change-Id: I598f775e877169b4a4d6b38ecff7a7ad8e727c5c