docs, Wrapper: Add dynamic sample code generation with controls
When a component demo is configurable, generate the code that would
result in the same output as the current controls state, and allow the
user to view and copy that. The `<cdx-demo-wrapper>` component must
receive a `show-generated-code` boolean property, and the name of
the component to use in the code is retrieved from the current
page title.

Uses the `prismjs` module for syntax highlighting, the highlighting of
the normal code displays is from Vitepress and could not be easily
reused.

Bug: T295184
Co-Authored-by: Catrope <roan@wikimedia.org>
Change-Id: I9f1c3d724eb69e6740ae163088cea97d6b7233ea