docs: Add button examples with icons
The Button docs didn't have any examples of buttons with icons at all.
There was one example of a progressive button with an icon, but it was
in the Icon docs.

Add a comprehensive set of examples to the Button docs, showing buttons
with an icon and text, as well as icon-only buttons, with different
actions and types. Also add an example of a quiet button, which was
missing.

Remove the example in the Icon docs, and instead link to the Button
docs.

Bug: T303982
Change-Id: If8f357a76136595122fafd4f5843bf58828785bf