Tabs: Introduce Tab and Tabs components, useIntersectionObserver
Provides the user with a <Tabs> component which must contain one
or more <Tab> components in its default slot. Each <Tab> component
can in turn contain arbitrary markup (including another nested
tabbed layout).

Semantically, the content of each Tab is presented as an HTML5
<section> tag with a unique ID. The Tabs component generates a
header list which is an unordered list of anchor links to the
same IDs. Appropriate ARIA roles like "tablist", "tabpanel",
etc. are used where necessary. The markup and accessibility
patterns are based on Heydon Pickering's "Tabbed Interfaces"
guide: https://inclusive-components.design/tabbed-interfaces/

In terms of the Vue.js API, the Tabs component uses a v-model
binding from the parent scope to track the currently active
tab. During setup, the props of child Tab components are
accessed (with useSlots) so that a data structure can be built
that contains all data for all tabs in a given layout. This is
used by the parent Tabs component to generate the header row,
handle keyboard navigation, and enable programmatic selection.

Each Tab component must have a unique "name" property; this can
be used as a key to access its data from the Tabs component.

Finally, the provide/inject API is used to provide the tabs
data to all child Tab components within a layout.

There are two available stylings for this component, "framed"
and frameless. This necessitates two sets of style rules for
many elements as appearance can vary significantly based on
which style is being used.

Regardless of the style, responsive design is achieved using
Flexbox to arrange tab labels into a horizontal row. This
row becomes scrollable on the horizontal axis when the labels
overflow the available space. To indicate whether more tabs
are available to the user by scrolling, a gradient indicator
will appear at the beginning and/or end of the tab header row.
The visibility of gradient indicators is determined using the
IntersectionObserver API.

A new composable, useIntersectionObserver, has been introduced
to streamline this. Currently the observers created by this
composable are very simple, they simply flip a boolean
ref to true or false based on whether the element they are
attached to is currently intersecting the viewport or not.

Bug: T303321
Change-Id: I556deefcd1422e5a4116ad760bb89bbd46d75e50