Menu: Change footer slot to no-results
This narrows the use case of this slot down to only no results
text, and moves it above the menu items. This is useful in cases
like TypeaheadSearch where an interactive menu item might be added
and displayed even when there are no actual menu items to show.

The Menu component handles the logic of when to show the no
results slot: when the slot has content and there are zero menu
items. A prop, showNoResultsSlot, has also been added to give
components that use Menu a way to override this default behavior.

Bug: T304457
Change-Id: I17fc934e3c6fd30b31418032cbd1b8f20eb373ea