MenuItem: Support language attributes
This patch adds a `language` property to the MenuItemData type,
which can contain language attributes for each text property
of a MenuItem (label, match, and description). It also adds the
`match` property, which is displayed after the label and will
most often be used to represent the text matching the search
query if it isn't the label or description (e.g. a Wikidata
alias).

Each text node of a MenuItem is now wrapped in a <bdi>
element to prevent surrounding text from influencing
its direction and vice versa.

MenuItem demos have been added to demonstrate the `match`
property and the language attribute support, and the
TypeaheadSearch demo for Wikidata has been updated to use
this new feature.

Bug: T305091
Change-Id: I6bf4e65ca052cc2e02a7cef700a31672bb0d6c5f