tokens: Add token type to JSON attributes
The token type is "base", "component" or "theme". This will help us
distinguish between these types of tokens in the future, for example
for T297923.

This implementation detects the type of the token based on the file
name, using the following patterns:
- If the file name starts with theme-, it's a theme token
- If the file name starts with codex-components, it's a component token
- If the file name starts with codex-base, it's a base token

This should be future proof for if we decide to spread base or component
tokens across multiple files, as long as we use a naming convention like
codex-components-button.json, codex-components-menu.json, etc. (We can
also choose a different convention and update this code.)

Change-Id: Ia3aa3706c6e3d226e2997c6d231ba65bc8e2a59b