docs: Unbreak navigating away from component pages with generated code
If a component page had generated code on it, navigating away from it
would break. I think this happened because the navigation process caused
the computed property for the component name to be recomputed, which
then called useRoute() while in the middle of a navigation action.

Fixed by calling useRoute() once in setup and reusing the result, rather
than calling it every time autoCodeTag is recomputed.

Bug: T305211
Change-Id: I1704be84ff2f9db474008e587c200a94e7dfe9db