build: Use rtlcss to generate codex.style-rtl.css, by running Vite twice
Replace our hacks built on postcss-rtlcss and a custom Vite plugin with
something much simpler:
- Run Vite twice, once with --mode=rtl and once without
- In RTL mode, write to codex.style-rtl.css instead of codex.style.css,
  and use rtlcss to flip the styles
- This wastefully builds the .js files twice, but it's still pretty fast
- In serve mode (npm run dev), continue to use postcss-rtlcss as before

This approach is a little clunky in that we have to run Vite twice, but
it's much easier to understand and reason about than the previous
approach.

Also configure esbuild so that its CSS minifier doesn't try to minify
'top: X; left: X; right: X; bottom: X;' to 'inset: X;'. We can't
do this because we support browsers that don't support inset (like
Safari 11-14). Before this change, this issue was avoided by accident:
postcss-rtlcss put the left/right properties in different CSS rules than
top/bottom, and then the split-bidi plugin merged them back together
post-minification, so the minifier never actually saw a rule that could
be minified to use the 'inset' property.

Bug: T304020
Change-Id: I698c30373ffec9b4415128b432bb5cf795cd2099