tokens: Use correct color in 'modifier-base80-translucent'
The code comment pointed in the right direction, but the color itself
was wrongly copying 'modifier-base90-translucent' over.
These colors have quite some history and are currently not-working
in WikimediaUI Base itself due to a broken calculation in Less.
Impacting quiet neutral buttons not to receive background color
changes on `:hover` and `:active`.
Will fix in WikimediaUI Base separately.

Bug: T305402
Change-Id: I78225724d7faca77a0c90b0efb4eb077f43c5d82