useModelWrapper: Support typed event parameters
Right now we don't have type information on any of our event parameters.
But in the future we might, or consumers of Codex might. In that case,
useModelWrapper() should ensure type safety by checking that the type
accepted by the update event is compatible with the type of the model
value prop.

Bug: T305500
Change-Id: I70a5a71c9f9c590c4e4380bcc2ccfd51f54523a9