docs: Ensure generated code samples can handle self-closing tags
The code originally intended to account for this, but when
slotValues.default is undefined, escapeHtml( slotValues.default )
will evaluate to the string 'undefined'. This patch adds an extra
conditional to check for default slot contents before running
that code, which made the logic complex enough to spell it out
via if statements rather than using a two-level ternary.

Bug: T305554
Change-Id: I36da25f104fdd369af9273ebfed57cc3b581656f