Tabs: Fix broken header hover styles in Chrome
Chrome (but not Fx or Safari) seemed to have some issues with hover
styles not being applied correctly when the rules were defined on
the wrapper `li` elements as opposed to the inner `a` elements.

See https://phabricator.wikimedia.org/T303321#7834644 for details
on the problem.

Moving the relevant `color` styles so they target the `[ role='tab' ]`
carrying `a` children elements directly fixes the issue.

Also
- moving styles to enabled/disabled selectors to contain scope
  and prevent unnecessary duplication of styles
- replacing `a` elements with `[ role='tab' ]` selectors to become
  element agnostic and align to Codex CSS guidelines.
- nesting the styles more strictly in parent-child hierarchy
- removing the "link unset" mixin as unnecessary
- updating the code comments for better orientation

Note, that with this patch
- the border-bottom transition is starting while the color transition
  is not. The border will be moved into the `a` element in a consecutive
  patch.
- the hover styles are out of sync due to the
  `transition` only be applied on the `.cdx-tabs__list__item` and it's
  not inherited in the child `a` element. Transition there will be
  established in consecutive patch if necessary with above.

Bug: T303321
Co-Authored-By: Volker E. <volker.e@wikimedia.org>
Change-Id: Ifdc9167d50393d9570c3510d069f254a3d27625d