Lookup: Use pending and focus states to decide whether to open the menu
Instead of looking at the text in the menu and figuring out whether it
matches the selected item, only open the menu on a menuItems change if
we're in the pending state. This indicates that the menuItems change was
in response to a new-input event.

Also don't open the menu when the input is not focused. TypeaheadSearch
already uses the same logic.

These changes fix the strange behavior where changing the menuItems prop
can cause the menu to be opened even if the Lookup isn't focused,
including the case where the user clicks out of the Lookup during the
time that the API request is pending. They also fix the bug where
removing the selected item from the menuItems prop in response to it
being selected would reopen the menu (effectively making the menu not
close on selection).

Also update the documentation to emphasize that the menuItems prop must
be updated after each new-input event, to prevent bugs where the Lookup
stays pending forever.

Bug: T304577
Change-Id: Ia5c76192f9f0af9813dfdb82fa42b6a95ac31fda