docs: Allow configuring icon properties and generate correct code
Add a new type of property configuration, 'icon', for properties where
the value is meant to be an Icon object. Show the user a lookup to
allow choosing what icon to use, where the values in the lookup are
the icon names. Then, in Wrapper.vue, convert the icon names to
actual icon objects for passing to the relevant wrapped component.

When a configurable demo has icon properties, a note is added to
the configuration table explaining that in addition to passing the
icon object to the component, the actual icon needs to be imported.

Additional changes:
* codegen.ts generateProps() now correctly uses kebab-case for
property names

* new helper class BoundProp is added to types for a property where
the value that is used for rendering is a string, but the value that
is actually passed to the component is not, to tell generateProps()
to use v-bind syntax for it

* codex-docs package jest configuration updated to handle
snapshots, with the code for snapshotResolver.js copied from
the file by the same name in the main codex package

* the configurable TextInput demo was updated to use the new icon
property types rather than boolean toggles for using specific icons,
and sample code is now generated for that demo

Bug: T306581
Change-Id: Ia7b6a64f181069fd5629383f361f2bc33599425f