Tabs: Add scroll buttons
Add buttons at the start and end of the tabs list that, when clicked,
scroll the tabs list 100px to the left/right. Only show each button when
there's room to scroll in that direction. Absolutely position the
buttons on top of the tabs list.

Because the arrow buttons are on top of the tabs list, using native
.scrollIntoView() for making the active tab visible doesn't fully work:
part of the tab is scrolled "into view" but underneath the arrow button.
To fix this, replace .scrollIntoView() with custom scrolling math that
takes the width of the arrow buttons into account.

When the scroll buttons are clicked, scroll the next partially invisible
tab into full view; unless that would scroll less than 25% of the width
of the tabs list, then scroll the next fully invisible into full view.

Smaller changes:
- Attach the gradient to the scroll buttons, instead of to the tabs list
- Wrap the prev/next buttons in divs with a solid background, to prevent
  the underlying tabs from shining through when the buttons become
  semi-transparent on hover
- Stretch the buttons to the full height of the tabs list
- Add an invisible focusHolder div, and focus it when the prev/next
  buttons are clicked. This ensures that we can listen to arrow key
  events without causing focus styling to appear.
- Remove @focus.prevent on the tabs, it doesn't do anything because the
  focus event is not preventable
- Remove observerAvailable, we don't need it anymore now that the
  gradient classes are gone
- Remove useless justify-content: space-between; rule on the tabs header

Bug: T307491
Change-Id: Ic1e51f948da09d238f3fb511c6c0f76608e669d2