tokens: Add `@position-offset-border-width-base`
`position-offset` is meant to be used as shorthand for top, right,
bottom, and left values. Instead of continuing to use a separate
token for `margin-offset`, this patch combines them in favor of
extending `position-offset` for use with margins as well. This
token is currently used to apply a negative border-width-base in
several instances of overlapping borders (a dropdown menu with its
associated input, a text input + button combination in SearchInput).

Bug: T307494
Change-Id: Ia576928bbc4ec6424a8da8e0c78067b9c3b4f1d2