ProgressBar: Set height on root element
Previously, the height of the progress bar was set on the "bar"
element within, while the border was set on the root element. This
caused the progress bar to be taller than expected (since the
`box-sizing: border-box` rule on the root element did not have the
intended effet.

This patch moves the height to the root element, which will take
into account the border size.

Bug: T295177
Change-Id: Id69fb31fe2b1a424ce24ec08d92e94690667b319