useIntersectionObserver: Make reactive to templateRef changing
When the templateRef passed to useIntersectionObserver changes, stop
observing the old element and start observing the new one.

This comes up in the Tabs demo when the "Swap out tabs" button is
clicked. The firstTabLabel and lastTabLabel template refs update to
point to the new tabs, but their IntersectionObservers didn't, so the
end gradient would stay even when you scrolled all the way to the end.

Change-Id: Ifbe44fe8e629e373840f462c7c5fbceed43067d0