docs: Allow configuring icons used as slot contents
Add a new type of configuration, 'slot-icon', for changing an icon
that is used as part of slot contents. The name of the control
is the relevant slot name followed by '-icon', and in the generated
sample code the icon will be added to that slot before any text.
For the actual displayed version, the icon name is passed in the
`slotValues` object under the same key it was configured with,
and then rendered using the new CdxDemoSlotIcon component,
which takes the name of the icon chosen, if there is one, and
renders the CdxIcon component for the icon with the given name.

The code in Wrapper.vue for separating out the property and slot
values and generating the example code is pretty messy at this
point, and will be cleaned up in a subsequent commit.

Additional changes:
* CdxDemoSlotIcon is added to the vitepress theme as a
component that is available in all demos, like the demo wrapper

* new helper class EscapedSlotContent is added to types for a
slot where the value includes an icon, such that the `<cdx-icon>`
component will not be escaped when generating demo code

* Wrapper.vue now handles escaping the text content of a slot
that also includes an icon, and passing to codegen.ts an
EscapedSlotContent object so that the text is not double escaped
and the icon code is not escaped at all

* the configurable Button demo was updated to use the new icon
slot option, with the ability to configure an icon in the button's
content

Bug: T306581
Change-Id: I7b606967bbeadd19eab60f67edabb3f88859e14b