docs: move more code generation logic from Wrapper.vue to codegen.ts
Reduce the bloat in Wrapper.vue, and the need for exporting the
`slotValueStrings` computed value for testing, by moving the
computations into codegen.ts. Wrapper now provides the title
of the component demo, the default configuration, and the current
configuration, and codegen.ts handles the rest, converting the
component demo title to a `cdx-*` component name, filtering out
properties that match the default, and dealing with escaping or
not escaping the properties and slots with icons.

Additional changes:
* The types BoundProp, EscapedSlotContent, and PropValues were
removed from types.ts, and SlotValues no longer includes
EscapedSlotContent. Instead, the first three were added to
codegen.ts instead, since they are only needed there, and
codegen.ts also includes a type SlotValuesWithIcons for when
the slots can include EscapedSlotContent

* Rather than needing to convert the title of the component demo,
like 'ToggleButton', to the component name for the generated code,
like 'cdx-toggle-button', that logic is now handled by
generateVueTag()

* Documentation is added to existing (and new) methods in
codegen.ts

* Wrapper.vue processing of controlsWithValues into propValues
and slotValues is deduplicated with a helper computed value,
splitValues

Bug: T306581
Change-Id: Ifec8d92ff5bd756bb466305354477de996178628