docs: add generic configurable component for using v-model
For a number of components the configurable demo simply binds
the properties to an instance of the component, and passes the slots.
However, for components were there is also a model value to
provide, simply binding the computed properties from the demo
controls does not work. Instead of needing to create a dedicated
configurable component for each of these demos, which just passes
the properties it is given along with a v-model reference, create a
generic component that does the same thing, determining the
component being demonstrated based on the page title.

This replaces the dedicated configurable demos for text input,
toggle button, and toggle switch.

Change-Id: I84dfa7510b2847c61807acc9aaed3b95e610fc24