useIntersectionObserver: Don't observe elements before they're mounted
Trying to observe an element before the component was mounted caused
strange issues. When the element was visible at mount time, the observer
would falsely tell us it wasn't visible, and we'd only get the right
answer after the user scrolled the page or resized the window.

By waiting to observe the element until the component is mounted, we do
get correct information.

Change-Id: Id59e143322b11e429ef4ff4799afbcde6e6f06ed