Tabs: Apply correct borders on hover/active
We accidentally removed the bottom border on frameless tabs for the
hover and active states. Before that, this bottom border was the wrong
color. It should match the text color, not be the same as the bottom
border color for selected tabs.

Bring back this bottom border and make its color match the text color in
each state.

When a tab is selected, we don't apply a hover or active state, because
no meaningful interaction is possible with a tab that is already
selected.

Bug: T303321
Change-Id: If1f4b79f2c88fe4cb59b0269d8d0b7594f00b474