Tabs, styles: Consistently apply margins to <a> elements
Most of our margins were applied to the <li>s, except for the framed
tabs' last-child margin, which was on the <a>.

Instead, consistently set the margins on the <a>s. This causes no visual
change, but ensures the bounding rects of the <li>s include the margins,
which fixes an issue where using the arrow keys to switch to the first
tab wouldn't scroll its left margin into view.

Bug: T307491
Change-Id: I517f27e04dac4b452db53da50233e405d631d5ac