docs: separate 'default' property values from initial values to use
The 'default' should match what the component actually uses as a
default, to ensure correct code generation. Add a new property,
'initial', supported on all configuration control types, for controlling
the value that is used when the demo is first loaded or is reset.

Update the message demo to have 'Close' be the *initial* label for
the dismiss button, rather than the *default* label.

Note that this does not properly work for icon properties yet, though
that is not an issue for the moment because there are no icon
properties where the initial value is being set. Icons will be fixed in
the follow-up commit that makes use of them.

Bug: T308715
Change-Id: I550c0e61edd0923c67e37385242fde246f275f19